requirements.txt
streamlit
pandas
openpyxl
numpy
networkx
python-dateutil
razorpay
gspread
google-auth